algorithm for finding in and max temp in range

+++ app/src/main/java/com/kinectpro/whattowear/data/wear/model/TempSummary.kt
@@ -5,4 +5,4 @@ package com.kinectpro.whattowear.data.wear.model
  * @param minValue minimum from range
  * @param maxValue maximum from range
  */
-data class TempSummary(val minValue: Double, val maxValue: Double)
+data class TempSummary(val minValue: Float, val maxValue: Float)
